Добавлен метод, свойство и событие авторизации в Модель.

+++ CommonNet6/Interfaces/IRepository.cs
@@ -7,7 +7,7 @@ namespace Interfaces
     public interface IRepository<T>
     {
         ReadOnlyObservableCollection<T> GetCollection();
-        bool Delete(int id);
+        bool Delete(long id);
         T Clone(T t);
         T? Add(T t);
         T? Update(T t);
